Added Next100 and Prev100 to alerts and alarms (Alexei)
git-svn-id: svn://svn.zabbix.com/trunk@417 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ src/zabbix_sucker/zabbix_sucker.c
@@ -308,7 +308,7 @@ int	get_value_SNMPv1(double *result,char *result_str,DB_ITEM *item)
 			)
 			{
 				*result=*vars->val.integer;
-				sprintf(result_str,"%d",*vars->val.integer);
+				sprintf(result_str,"%ld",(long)*vars->val.integer);
 			}
 			else if(vars->type == ASN_OCTET_STR)
 			{
